+++ src/main/java/chess/domain/Board.java
@@ -1,6 +1,7 @@
-package chess;
+package chess.domain;
 
-import chess.piece.Piece;
+import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
