+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,6 +1,4 @@
-package chess.piece;
-
-import chess.Team;
+package chess.domain.piece;
 
 public abstract class Piece {
