+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public final class Queen extends chess.domain.piece.Piece {
+
+    public Queen(Team team) {
+        super(team);
+    }
+}
