+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public final class Rook extends chess.domain.piece.Piece {
+
+    public Rook(Team team) {
+        super(team);
+    }
+}
