+++ src/main/java/chess/domain/piece/Team.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain.piece;
 
 public enum Team {
     BLACK, WHITE;
