--- src/main/java/chess/piece/Queen.java
@@ -1,10 +1,0 @@
-package chess.piece;
-
-import chess.Team;
-
-public final class Queen extends Piece {
-
-    public Queen(Team team) {
-        super(team);
-    }
-}
