--- src/main/java/chess/piece/Rook.java
@@ -1,10 +1,0 @@
-package chess.piece;
-
-import chess.Team;
-
-public final class Rook extends Piece {
-
-    public Rook(Team team) {
-        super(team);
-    }
-}
