+++ src/test/java/chess/domain/piece/KingTest.java
@@ -1,8 +1,7 @@
-package chess.piece;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.Team;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -13,7 +12,7 @@ class KingTest {
     @CsvSource({"BLACK, true", "WHITE, false"})
     @DisplayName("해당 팀이 검정 팀인지 확인한다.")
     void isBlackTeamTest(Team team, boolean expected) {
-        King king = new King(team);
+        chess.domain.piece.King king = new chess.domain.piece.King(team);
 
         assertThat(king.isBlackTeam()).isEqualTo(expected);
     }
